Unificiation of Caching logic for MvxCachingFragmentCompatActivity.
