+++ Cirrious.MvvmCross.Droid.Support.Fragging/Cirrious.MvvmCross.Droid.Support.Fragging.csproj
@@ -80,6 +80,7 @@
     <Compile Include="Fragments\EventSource\MvxBaseFragmentAdapter.cs" />
     <Compile Include="Fragments\MvxBindingFragmentAdapter.cs" />
     <Compile Include="Fragments\MvxSimpleLayoutInflaterHolder.cs" />
+    <Compile Include="Cache\MvxActivityCachingProcessor.cs" />
     <Compile Include="MvxOwnedViewModelFragmentAttribute.cs" />
     <Compile Include="MvxOwnedViewModelFragmentAttributeExtensionMethods.cs" />
     <Compile Include="MvxCachingFragmentActivity.cs" />
